requirements.txt
paramiko
pysmb